+++ react/typings/sbdsefuat.checkout-simulation.d.ts
@@ -1,0 +1,1 @@
+declare module 'sbdsefuat.checkout-simulation'